somacor_cmms/backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-decouple>=3.6
Pillow>=9.0.0

